backend/requirements.txt
# ============================
# 上课摸鱼搭子 - 后端依赖清单
# ============================

# Web 框架
fastapi==0.115.6
uvicorn[standard]==0.34.0

# 跨域支持
python-multipart==0.0.20

# 环境变量
python-dotenv==1.0.1

# 本地 Ollama HTTP 客户端
httpx==0.28.1

# PPT 解析
python-pptx==1.0.2

# PDF 解析
pypdf>=4.0.0

# Word 文档解析
python-docx>=1.1.0

# 音频录制
pyaudio==0.2.14

# 历史兼容云端链路排障（仅 backend/test_llm.py 使用）
openai==1.58.1

# ASR - 阿里云百炼 DashScope
dashscope>=1.20.0

# ASR - 字节跳动 Seed-ASR (WebSocket 客户端)
websocket-client>=1.7.0

# WebSocket 支持 (FastAPI 自带，但需要 websockets 库)
websockets==14.1

# ASR - 本地免费语音识别 (Google 免费 API，无需密钥)
SpeechRecognition>=3.10.0

# 异步文件操作
aiofiles==24.1.0
